Library: move Analysis, Hidden and Missing into separate directories
