+++ CMakeLists.txt
@@ -888,8 +888,10 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/engine/sync/internalclock.cpp
   src/engine/sync/synccontrol.cpp
   src/errordialoghandler.cpp
-  src/library/analysisfeature.cpp
-  src/library/analysislibrarytablemodel.cpp
+  src/library/analysis/analysisfeature.cpp
+  src/library/analysis/analysislibrarytablemodel.cpp
+  src/library/analysis/dlganalysis.cpp
+  src/library/analysis/dlganalysis.ui
   src/library/autodj/autodjfeature.cpp
   src/library/autodj/autodjprocessor.cpp
   src/library/autodj/dlgautodj.cpp
@@ -923,14 +925,8 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/library/dao/settingsdao.cpp
   src/library/dao/trackdao.cpp
   src/library/dao/trackschema.cpp
-  src/library/dlganalysis.cpp
-  src/library/dlganalysis.ui
   src/library/dlgcoverartfullsize.cpp
   src/library/dlgcoverartfullsize.ui
-  src/library/dlghidden.cpp
-  src/library/dlghidden.ui
-  src/library/dlgmissing.cpp
-  src/library/dlgmissing.ui
   src/library/dlgtagfetcher.cpp
   src/library/dlgtagfetcher.ui
   src/library/dlgtrackinfo.cpp
@@ -942,7 +938,6 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/library/export/trackexportwizard.cpp
   src/library/export/trackexportworker.cpp
   src/library/externaltrackcollection.cpp
-  src/library/hiddentablemodel.cpp
   src/library/itunes/itunesdao.cpp
   src/library/itunes/itunesfeature.cpp
   src/library/itunes/itunesimporter.cpp
@@ -954,7 +949,12 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/library/libraryfeature.cpp
   src/library/librarytablemodel.cpp
   src/library/locationdelegate.cpp
-  src/library/missingtablemodel.cpp
+  src/library/missing_hidden/dlghidden.cpp
+  src/library/missing_hidden/dlghidden.ui
+  src/library/missing_hidden/dlgmissing.cpp
+  src/library/missing_hidden/dlgmissing.ui
+  src/library/missing_hidden/hiddentablemodel.cpp
+  src/library/missing_hidden/missingtablemodel.cpp
   src/library/mixxxlibraryfeature.cpp
   src/library/multilineeditdelegate.cpp
   src/library/parser.cpp
