+++ src/library/missing_hidden/dlghidden.cpp
@@ -1,10 +1,10 @@
-#include "library/dlghidden.h"
+#include "library/missing_hidden/dlghidden.h"
 
 #include <QItemSelection>
 
 #include "controllers/keyboard/keyboardeventfilter.h"
-#include "library/hiddentablemodel.h"
 #include "library/library.h"
+#include "library/missing_hidden/hiddentablemodel.h"
 #include "moc_dlghidden.cpp"
 #include "util/assert.h"
 #include "widget/wlibrary.h"
